the count of commits is now a u16. I think 65535 possible commits should bring us far enough.

+++ src/gitcommands.rs
@@ -1,6 +1,6 @@
 use std::process::Command;
 
-pub fn get_commit_count(path: &str) -> Option<usize> {
+pub fn get_commit_count(path: &str) -> Option<u16> {
     let output = Command::new("git")
         .current_dir(path)
         .arg("rev-list")
@@ -12,7 +12,7 @@ pub fn get_commit_count(path: &str) -> Option<usize> {
     let commit_count = String::from_utf8(output.stdout)
         .expect("Could not translate output from utf8");
     let commit_count = commit_count.trim();
-    let commit_count = commit_count.parse::<usize>();
+    let commit_count = commit_count.parse::<u16>();
 
     match commit_count {
         Ok(commit_count) => { Some(commit_count) }
